[Ref]: 영화 검색 handler 코드 분리
